requirements.txt
requests
googlemaps
numpy
ratelimit
ortools
pandas
python-dotenv
shapely
qrcode
reportlab
polars

